FIX: change datatype of document column to bigint

+++ db/migrate/20221005153854_create_users.rb
@@ -4,7 +4,7 @@ class CreateUsers < ActiveRecord::Migration[7.0]
       t.integer :user_type, default: 0
       t.string :name
       t.string :last_name
-      t.integer :document
+      t.bigint :document
       t.string :password_digest
       t.float :intake, default: 0
       t.integer :document_type, default: 0
